feat(new-market-overview): refactor /lender/components directory
