+++ src/app/[locale]/borrower/components/MlaModal/interface.ts
