+++ src/app/[locale]/borrower/create-market/components/Forms/ConfirmationForm/index.tsx
@@ -4,9 +4,9 @@ import { Token } from "@wildcatfi/wildcat-sdk"
 import { UseFormReturn } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 
+import { MlaModal } from "@/app/[locale]/borrower/components/MlaModal"
 import { usePreviewMlaFromForm } from "@/app/[locale]/borrower/hooks/mla/usePreviewMla"
 import { SignMlaFromFormInputs } from "@/app/[locale]/borrower/hooks/mla/useSignBorrowerMla"
-import { MlaModal } from "@/app/[locale]/lender/components/MlaModal"
 import { BorrowerProfile } from "@/app/api/profiles/interface"
 import BackArrow from "@/assets/icons/arrowLeft_icon.svg"
 import Info from "@/assets/icons/info_icon.svg"
