+++ src/app/[locale]/lender/components/ExploreSection/TrendingMarketsCarousel/TrendingMarketCardApr/index.tsx
