+++ src/app/[locale]/lender/components/ExploreSection/TrendingMarketsCarousel/TrendingMarketCardInflow/index.tsx
