+++ src/app/[locale]/lender/components/ExploreSection/TrendingMarketsCarousel/TrendingMarketCardInterestPaid/index.tsx
