+++ src/app/[locale]/lender/components/ExploreSection/TrendingMarketsCarousel/index.tsx
